Issue "Make sure input fields are not empty #16"
Following are the Modification performed for acheiving the functionality : 
a) If the textbox of "Hour" or "minute" is empty , then its respective varaible values are set to 0 ( in cookies ) along with endTime as "00:00". 
b) If the textbox of "StartTime" is empty , then startTime will be set as default time ( i.e. "00:00" ). 
Intuition behind taking default value is that , if in case values of "hour" and "minute" textbox are correct then , endTime will be shown by combination of defaultTime (00:00) with addition of respective hour and minute values present in textbox.
In case if either of "hour" and "minute" textbox value(s) are blank , then endTime will be displayed as "00:00".